4.22*: bloglist step10 - need to complet tests after auth

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({})
@@ -9,7 +11,13 @@ blogsRouter.get('/', async (request, response) => {
   
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    const user = await User.findById(body.userId)
+    const token = request.token //token set by missleware
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -26,14 +34,27 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
-    if (deletedBlog) {
-        response.status(204).end()
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
     }
-    else {
-        response.status(400).json( 
-            {error: `no blog with the following id: ${request.params.id}`})
+    const user = await User.findById(decodedToken.id)
+    const blogToDelete = await Blog.findById(request.params.id)
+    if (!blogToDelete) return response.status(400).json({error: `no blog with the following id: ${request.params.id}`})
+    
+    else if (blogToDelete.user.toString() === user.id.toString()){
+        const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
+        if (deletedBlog) {
+            user.blogs = user.blogs.filter(blog => blog.id !== deletedBlog.id)
+            response.status(204).end()
+        }
+        else {
+            response.status(400).json( 
+                {error: `no blog with the following id: ${request.params.id}`})
+        }
     }
+    
 })
 
 blogsRouter.put('/:id', async (request, response) => {
